corrected flickering
corrected flickering that happens when a new task is created.  All the the tasks were being received from the server and redrawn every time a task was created.  Now the new task is pushed, no data is received from the server, and only the new task is drawn on the screen.